[patch,microblaze]: Fix the calculation of high word in a long long 6. .4-bit
This patch will change the calculation of high word in a long long 64-bit.
Earlier to this patch the high word of long long word (0xF0000000ULL) is
coming to be 0xFFFFFFFF and low word is 0xF0000000. Instead the high word
should be 0x00000000 and the low word should be 0xF0000000. This patch
removes the condition of checking high word = 0 & low word < 0.
This check is not required for the correctness of calculating 32-bit high
and low words in a 64-bit long long.

Signed-off-by :Nagaraju Mekala <nmekala@xilix.com>
               Ajit Agarwal  <ajitkum@xilinx.com>

ChangeLog:
2016-03-01 Nagaraju Mekala <nmekala@xilix.com>
           Ajit Agarwal  <ajitkum@xilinx.com>

    *config/microblaze/microblaze.c (print_operand): Remove the condition of checking
    high word = 0 & low word < 0.
    *testsuite/gcc.target/microblaze/others/long.c: Add -O0 option.